refactored VectorMCMC class #8
